Extract scene lighting, intro pose, and shared props.
Move setupLighting and getIntroPose out of Scene.tsx and share SceneProps so the monolith is easier to navigate without behavior changes.
